+++ yggdrasil_decision_forests/utils/feature_importance.h
@@ -27,6 +27,15 @@
 namespace yggdrasil_decision_forests {
 namespace utils {
 
+struct ComputeFeatureImportanceOptions {
+  // Number of threads used for the computation. Should be >=1.
+  int num_threads = 6;
+
+  // Number of repetitions of the estimation. More repetitions increase the
+  // quality of the variable importance estimates. Should be >=1.
+  int num_rounds = 1;
+};
+
 // Computes and adds to the model permutation feature importances.
 absl::Status ComputePermutationFeatureImportance(
     const metric::proto::EvaluationResults& base_evaluation,
@@ -34,12 +43,13 @@ absl::Status ComputePermutationFeatureImportance(
         absl::StatusOr<absl::optional<metric::proto::EvaluationResults>>(
             const int feature_idx)>
         get_permutation_evaluation,
-    model::AbstractModel* model, int num_rounds = 1);
+    model::AbstractModel* model,
+    const ComputeFeatureImportanceOptions& options = {});
 
 // Computes and adds to the model permutation feature importances.
 absl::Status ComputePermutationFeatureImportance(
     const dataset::VerticalDataset& dataset, model::AbstractModel* model,
-    int num_rounds = 1);
+    const ComputeFeatureImportanceOptions& options = {});
 
 // Builds a copy of the dataset with the values of the columns in
 // "shuffle_column_idxs" are shuffled randomly.